+++ scheme/GodunovImplemV0.cpp
@@ -287,7 +287,7 @@ GodunovImplemV0<dim, device_t>::compute_fluxes_and_store_in_owned_and_ghosts(rea
 
   // reshape flux to be a flux array in given direction
   auto flux_block_sizes = m_Q_ghosted.block_size();
-  flux_block_sizes[direction]++;
+  flux_block_sizes[static_cast<size_t>(direction)]++;
   m_Fluxes.reshape(flux_block_sizes);
 
   // compute fluxes and update all quadrants in a group of quadrants
@@ -328,7 +328,7 @@ GodunovImplemV0<dim, device_t>::compute_viscous_fluxes_and_store_in_owned_and_gh
 
   // reshape flux to be a flux array in given direction
   auto flux_block_sizes = m_Q_ghosted.block_size();
-  flux_block_sizes[direction]++;
+  flux_block_sizes[static_cast<size_t>(direction)]++;
   m_Fluxes.reshape(flux_block_sizes);
 
   // compute fluxes and update all quadrants in a group of quadrants
@@ -365,7 +365,7 @@ GodunovImplemV0<dim, device_t>::read_fluxes_and_update_in_owned(DataArrayBlock_t
   // check flux array sizes
   {
     [[maybe_unused]] auto flux_block_sizes = m_Q_ghosted.block_size();
-    flux_block_sizes[direction]++;
+    flux_block_sizes[static_cast<size_t>(direction)]++;
     assertm(flux_block_sizes == m_Fluxes.shape(), "Flux array has incompatible shape.");
   }
